WebProject/requirements.txt
python
Flask
Flask-Login
Flask-WTF
Pillow
SQLAlchemy
SQLAlchemy-serializer
WTForms
Werkzeug
pip